Fix: SkillParser & RegionSkills
- Now all skills are being parse correctly.
- Adding more skill type checks.
- Adding GM skills type.
- Now Skill Region are being cast correctly. (WIP)
- Correct implementation of skill Levels.
- Minor other fixes.

+++ MapleServer2/Database/Classes/DatabaseSkillTab.cs
@@ -25,7 +25,7 @@ namespace MapleServer2.Database.Classes
             List<SkillTab> skillTabs = new List<SkillTab>();
             foreach (dynamic item in skillTabsResult)
             {
-                skillTabs.Add(new SkillTab(item.Name, jobId, item.TabId, item.Uid, JsonConvert.DeserializeObject<Dictionary<int, int>>(item.SkillLevels)));
+                skillTabs.Add(new SkillTab(item.Name, jobId, item.TabId, item.Uid, JsonConvert.DeserializeObject<Dictionary<int, short>>(item.SkillLevels)));
             }
             return skillTabs;
         }
